Move medium compute to medium_physics
