+++ core/specfem/assembly/assembly/impl/helper.hpp
@@ -1,11 +1,11 @@
 #pragma once
 
 #include "enumerations/interface.hpp"
-#include "medium/medium.hpp"
 #include "specfem/assembly/assembly.hpp"
 #include "specfem/assembly/assembly/impl/helper.hpp"
 #include "specfem/chunk_element.hpp"
 #include "specfem/execution.hpp"
+#include "specfem/medium_physics.hpp"
 #include "specfem/parallel_configuration.hpp"
 #include "specfem/point.hpp"
 #include <Kokkos_Core.hpp>
