+++ core/specfem/kokkos_kernels/impl/compute_seismogram.tpp
@@ -9,7 +9,7 @@
 #include "enumerations/medium.hpp"
 #include "enumerations/wavefield.hpp"
 #include "specfem/execution.hpp"
-#include "medium/compute_wavefield.hpp"
+#include "specfem/medium_physics.hpp"
 #include "specfem/parallel_configuration.hpp"
 #include "specfem/assembly.hpp"
 #include "specfem/point.hpp"
