+++ core/specfem/kokkos_kernels/impl/compute_stiffness_interaction.tpp
@@ -7,10 +7,7 @@
 #include "enumerations/medium.hpp"
 #include "enumerations/wavefield.hpp"
 #include "specfem/execution.hpp"
-#include "medium/compute_cosserat_couple_stress.hpp"
-#include "medium/compute_cosserat_stress.hpp"
-#include "medium/compute_damping_force.hpp"
-#include "medium/compute_stress.hpp"
+#include "specfem/medium_physics.hpp"
 #include "specfem/parallel_configuration.hpp"
 #include "specfem/quadrature.hpp"
 #include "specfem/assembly.hpp"
