+++ core/specfem/medium_physics.hpp
@@ -16,6 +16,12 @@
  */
 namespace specfem::medium {}
 
-#include "compute_frechet_derivatives.hpp"
-#include "compute_stress.hpp"
-#include "compute_wavefield.hpp"
+#include "medium_physics/compute_cosserat_couple_stress.hpp"
+#include "medium_physics/compute_cosserat_stress.hpp"
+#include "medium_physics/compute_coupling.hpp"
+#include "medium_physics/compute_damping_force.hpp"
+#include "medium_physics/compute_frechet_derivatives.hpp"
+#include "medium_physics/compute_mass_matrix.hpp"
+#include "medium_physics/compute_source.hpp"
+#include "medium_physics/compute_stress.hpp"
+#include "medium_physics/compute_wavefield.hpp"
