+++ core/specfem/medium_physics/compute_cosserat_couple_stress.hpp
