+++ core/specfem/medium_physics/compute_cosserat_stress.hpp
