+++ core/specfem/medium_physics/compute_damping_force.hpp
