+++ core/specfem/medium_physics/compute_mass_matrix.hpp
@@ -1,13 +1,13 @@
 #pragma once
 
-#include "dim2/acoustic/isotropic/mass_matrix.tpp"
-#include "dim2/elastic/anisotropic/mass_matrix.tpp"
-#include "dim2/elastic/isotropic/mass_matrix.tpp"
-#include "dim2/elastic/isotropic_cosserat/mass_matrix.tpp"
-#include "dim2/poroelastic/isotropic/mass_matrix.tpp"
-#include "dim3/elastic/isotropic/mass_matrix.tpp"
 #include "enumerations/dimension.hpp"
 #include "enumerations/medium.hpp"
+#include "medium/dim2/acoustic/isotropic/mass_matrix.tpp"
+#include "medium/dim2/elastic/anisotropic/mass_matrix.tpp"
+#include "medium/dim2/elastic/isotropic/mass_matrix.tpp"
+#include "medium/dim2/elastic/isotropic_cosserat/mass_matrix.tpp"
+#include "medium/dim2/poroelastic/isotropic/mass_matrix.tpp"
+#include "medium/dim3/elastic/isotropic/mass_matrix.tpp"
 #include "specfem/macros.hpp"
 #include "specfem/point.hpp"
 #include "specfem_setup.hpp"
