+++ core/specfem/medium_physics/compute_source.hpp
@@ -1,11 +1,11 @@
 #pragma once
 
-#include "dim2/acoustic/isotropic/source.hpp"
-#include "dim2/elastic/anisotropic/source.hpp"
-#include "dim2/elastic/isotropic/source.hpp"
-#include "dim2/elastic/isotropic_cosserat/source.hpp"
-#include "dim2/poroelastic/isotropic/source.hpp"
-#include "dim3/elastic/isotropic/source.hpp"
+#include "medium/dim2/acoustic/isotropic/source.hpp"
+#include "medium/dim2/elastic/anisotropic/source.hpp"
+#include "medium/dim2/elastic/isotropic/source.hpp"
+#include "medium/dim2/elastic/isotropic_cosserat/source.hpp"
+#include "medium/dim2/poroelastic/isotropic/source.hpp"
+#include "medium/dim3/elastic/isotropic/source.hpp"
 #include "specfem/data_access.hpp"
 #include <Kokkos_Core.hpp>
