+++ core/specfem/medium_physics/compute_stress.hpp
@@ -1,11 +1,11 @@
 #pragma once
 
-#include "dim2/acoustic/isotropic/stress.hpp"
-#include "dim2/elastic/anisotropic/stress.hpp"
-#include "dim2/elastic/isotropic/stress.hpp"
-#include "dim2/elastic/isotropic_cosserat/stress.hpp"
-#include "dim2/poroelastic/isotropic/stress.hpp"
-#include "dim3/elastic/isotropic/stress.hpp"
+#include "medium/dim2/acoustic/isotropic/stress.hpp"
+#include "medium/dim2/elastic/anisotropic/stress.hpp"
+#include "medium/dim2/elastic/isotropic/stress.hpp"
+#include "medium/dim2/elastic/isotropic_cosserat/stress.hpp"
+#include "medium/dim2/poroelastic/isotropic/stress.hpp"
+#include "medium/dim3/elastic/isotropic/stress.hpp"
 #include "specfem/data_access.hpp"
 #include <Kokkos_Core.hpp>
