+++ core/specfem/medium_physics/compute_wavefield.hpp
@@ -1,11 +1,11 @@
 #pragma once
 
-#include "dim2/acoustic/isotropic/wavefield.hpp"
-#include "dim2/elastic/anisotropic/wavefield.hpp"
-#include "dim2/elastic/isotropic/wavefield.hpp"
-#include "dim2/elastic/isotropic_cosserat/wavefield.hpp"
-#include "dim2/poroelastic/isotropic/wavefield.hpp"
-#include "dim3/elastic/isotropic/wavefield.hpp"
+#include "medium/dim2/acoustic/isotropic/wavefield.hpp"
+#include "medium/dim2/elastic/anisotropic/wavefield.hpp"
+#include "medium/dim2/elastic/isotropic/wavefield.hpp"
+#include "medium/dim2/elastic/isotropic_cosserat/wavefield.hpp"
+#include "medium/dim2/poroelastic/isotropic/wavefield.hpp"
+#include "medium/dim3/elastic/isotropic/wavefield.hpp"
 #include <Kokkos_Core.hpp>
 
 namespace specfem {
