+++ include/medium/dim2/acoustic/isotropic/wavefield.hpp
@@ -2,9 +2,10 @@
 
 #include "enumerations/dimension.hpp"
 #include "enumerations/medium.hpp"
-#include "medium/compute_stress.hpp"
 #include "specfem/algorithms.hpp"
+#include "specfem/assembly.hpp"
 #include "specfem/macros.hpp"
+#include "specfem/medium_physics.hpp"
 #include "specfem/point.hpp"
 #include <Kokkos_Core.hpp>
