+++ include/medium/dim2/elastic/anisotropic/wavefield.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "enumerations/medium.hpp"
-#include "medium/compute_stress.hpp"
 #include "specfem/algorithms.hpp"
+#include "specfem/medium_physics.hpp"
 #include "specfem/point.hpp"
 #include <Kokkos_Core.hpp>
