+++ include/medium/dim2/elastic/isotropic/wavefield.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
 #include "enumerations/medium.hpp"
-#include "medium/compute_stress.hpp"
 #include "specfem/algorithms.hpp"
 #include "specfem/macros.hpp"
+#include "specfem/medium_physics.hpp"
 #include "specfem/point.hpp"
 #include <Kokkos_Core.hpp>
