+++ tests/unit-tests/medium/mass_matrix/dim2/acoustic.cpp
@@ -1,4 +1,4 @@
-#include "medium/compute_mass_matrix.hpp"
+#include "specfem/medium_physics.hpp"
 #include "specfem/point.hpp"
 #include <gtest/gtest.h>
 #include <sstream>
